+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [0.0.4] - 2025-08-12
+### Added
+- Support for WDePOS on iOS and Android.
+- Updated generateNonce method to use WDePOS.
+
+### Fixed
+- Fixed integration with FlutterMethodChannel.